added counter to know location for JE-PLG
